[test] Service, Repository 테스트 작성중

+++ src/main/java/org/crops/fitserver/domain/user/dto/UserInfoDto.java
@@ -32,7 +32,7 @@ public class UserInfoDto {
   private Boolean isOpenProfile;
   private String portfolioUrl;
   private Integer projectCount;
-  private Integer activityHour;
+  private Short activityHour;
   private String introduce;
   private List<Link> linkList;
   private Long positionId;
